Add formatted output for the `LauncherResponse` type
This changes an error message from
```
Expected "(int)response" == "(int)LauncherResponse::kSuccess" but was 69 vs 83.
```
to
```
Expected "response" == "LauncherResponse::kSuccess" but was LauncherResponse::kError vs LauncherResponse::kSuccess.
```

Bug: b/446765997

+++ base/cvd/cuttlefish/host/libs/command_util/util.cc
@@ -164,7 +164,7 @@ Result<void> RunLauncherAction(SharedFD monitor_socket, LauncherAction action,
   LauncherResponse response;
   CF_EXPECT(ReadExactBinaryResult(monitor_socket, &response),
             "Error reading LauncherResponse");
-  CF_EXPECT_EQ((int)response, (int)LauncherResponse::kSuccess);
+  CF_EXPECT_EQ(response, LauncherResponse::kSuccess);
   return {};
 }
 
@@ -194,7 +194,7 @@ Result<void> RunLauncherAction(
   LauncherResponse response;
   CF_EXPECT(ReadExactBinaryResult(monitor_socket, &response),
             "Error reading LauncherResponse");
-  CF_EXPECT_EQ((int)response, (int)LauncherResponse::kSuccess);
+  CF_EXPECT_EQ(response, LauncherResponse::kSuccess);
   return {};
 }
